transaction_data - fetch() methods are stubs so far.
all methods working except fetch() ones. waiting to make changes to libbitcoin
first in its serializer. I chose to make buckets a fixed compile time value
set in the code. we can later make it configurable if there's a need but probably
best if I choose this value for when new databases are initialized.