Call block_organizer_.stop() pre critical section
This seems to be working for me, without causing any problem with database integrity from discarding pending queued validation work, to signal stop before the validation_mutex_ lock is requested avoids a potentially very long wait for an existing validation sequence to finish and for its mutex lock to be released. Looks like block_organizer::stop() merely sets stop flags, and downloader_subscriber_->stop() does not seem to require validation_mutex_ to be locked so it looks safe to move this here.